requirements.txt
pygame
pillow
ttkbootstrap